Update outdated docs
