+++ src/load/loading_screen.rs
@@ -13,6 +13,10 @@
 //! [`Task`]: ../struct.Task.html
 //! [`LoadingScreen`]: trait.LoadingScreen.html
 //! [`ProgressBar`]: struct.ProgressBar.html
+mod progress_bar;
+
+pub use progress_bar::ProgressBar;
+
 use super::{Progress, Task};
 use crate::graphics;
 use crate::Result;
@@ -21,65 +25,8 @@ use crate::Result;
 /// to the user.
 ///
 /// # Usage
-/// If you have a [`LoadingScreen`], you can use it in your [`Game::new`] method
-/// easily. Let's say we want to use the [`ProgressBar`] loading screen in our
-/// game:
-///
-/// ```
-/// use coffee::{Game, Result};
-/// use coffee::load::{Task, Join, LoadingScreen};
-/// use coffee::load::loading_screen::ProgressBar;
-/// use coffee::graphics::Window;
-/// # use coffee::Timer;
-/// # use coffee::graphics::{Frame, Gpu};
-/// #
-/// # struct State;
-/// # impl State {
-/// # fn load() -> Task<State> { Task::new(|| State) }
-/// # }
-/// # struct View;
-/// # impl View {
-/// # fn load() -> Task<View> { Task::new(|| View) }
-/// # }
-/// # struct Input;
-/// # impl Input {
-/// # fn new() -> Input { Input }
-/// # }
-///
-/// struct MyGame {
-///     state: State,
-///     // ...
-/// }
-///
-/// impl Game for MyGame {
-/// #   type View = View;
-/// #   type Input = Input;
-/// #
-/// #   const TICKS_PER_SECOND: u16 = 60;
-/// #
-///     // ...
-///
-///     fn new(window: &mut Window) -> Result<(MyGame, View, Input)> {
-///         let load =
-///             (
-///                 Task::stage("Loading state...", State::load()),
-///                 Task::stage("Loading assets...", View::load()),
-///             )
-///                 .join();
-///
-///         // Create the loading screen and use `run`
-///         let mut progress_bar = ProgressBar::new(window.gpu());
-///         let (state, view) = progress_bar.run(load, window)?;
-///
-///         Ok((MyGame { state }, view, Input::new()))
-///     }
-///
-///     // ...
-///     # fn update(&mut self, _view: &View, window: &Window) {}
-///     # fn draw(&self, _view: &mut Self::View, _frame: &mut Frame,
-///     #         _timer: &Timer) {}
-/// }
-/// ```
+/// If you have a [`LoadingScreen`], simply set it as your [`Game::LoadingScreen`]
+/// associated type. Coffee will automatically use it when your game starts!
 ///
 /// # Future plans
 /// As of now, Coffee only ships with the [`ProgressBar`] loading screen. In the
@@ -91,7 +38,7 @@ use crate::Result;
 /// [`Task`]: ../struct.Task.html
 /// [`LoadingScreen`]: trait.LoadingScreen.html
 /// [`ProgressBar`]: struct.ProgressBar.html
-/// [`Game::new`]: ../../trait.Game.html#tymethod.new
+/// [`Game::LoadingScreen`]: ../../trait.Game.html#associatedtype.LoadingScreen
 /// [create an issue]: https://github.com/hecrj/coffee/issues
 /// [open a pull request]: https://github.com/hecrj/coffee/pulls
 pub trait LoadingScreen {
@@ -102,15 +49,11 @@ pub trait LoadingScreen {
     /// React to task progress.
     ///
     /// You should provide feedback to the user here. You can draw on the given
-    /// [`Window`], like in [`Game::draw`].
+    /// [`Frame`], like in [`Game::draw`].
     ///
-    /// [`Window`]: ../../graphics/struct.Window.html
+    /// [`Frame`]: ../../graphics/struct.Frame.html
     /// [`Game::draw`]: ../../trait.Game.html#tymethod.draw
-    fn on_progress(
-        &mut self,
-        progress: &Progress,
-        window: &mut graphics::Window,
-    );
+    fn on_progress(&mut self, progress: &Progress, frame: &mut graphics::Frame);
 
     /// Run the loading screen with a task and obtain its result.
     ///
@@ -122,83 +65,21 @@ pub trait LoadingScreen {
         window: &mut graphics::Window,
     ) -> Result<T> {
         task.run(window, |progress, window| {
-            self.on_progress(progress, window);
+            self.on_progress(progress, &mut window.frame());
             window.swap_buffers();
         })
     }
 }
 
-/// A simple loading screen showing a progress bar and the current stage.
-///
-/// ![The ProgressBar loading screen][progress_bar]
-///
-/// See [`LoadingScreen`] for a detailed example on how to use it.
-///
-/// [progress_bar]: https://github.com/hecrj/coffee/blob/e079e7205a53f92ac6614382b5cdd250fed64a98/images/loading_screen/progress_bar.png?raw=true
-/// [`LoadingScreen`]: trait.LoadingScreen.html
-#[allow(missing_debug_implementations)]
-pub struct ProgressBar {
-    font: graphics::Font,
-    pencil: graphics::Image,
-}
-
-impl LoadingScreen for ProgressBar {
-    /// Create the loading screen.
-    fn new(gpu: &mut graphics::Gpu) -> Result<Self> {
-        Ok(Self {
-            font: graphics::Font::from_bytes(gpu, graphics::Font::DEFAULT)?,
-            pencil: graphics::Image::from_colors(
-                gpu,
-                &[graphics::Color::WHITE],
-            )?,
-        })
+impl LoadingScreen for () {
+    fn new(_gpu: &mut graphics::Gpu) -> Result<Self> {
+        Ok(())
     }
 
     fn on_progress(
         &mut self,
-        progress: &Progress,
-        window: &mut graphics::Window,
+        _progress: &Progress,
+        _frame: &mut graphics::Frame,
     ) {
-        let mut frame = window.frame();
-
-        frame.clear(graphics::Color::BLACK);
-
-        self.pencil.draw(
-            graphics::Quad {
-                position: graphics::Point::new(
-                    50.0,
-                    frame.height() / 2.0 - 25.0,
-                ),
-                size: (
-                    (frame.width() - 100.0) * (progress.percentage() / 100.0),
-                    50.0,
-                ),
-                ..Default::default()
-            },
-            &mut frame.as_target(),
-        );
-
-        if let Some(stage) = progress.stage() {
-            self.font.add(graphics::Text {
-                content: stage,
-                position: graphics::Point::new(
-                    50.0,
-                    frame.height() / 2.0 - 80.0,
-                ),
-                size: 30.0,
-                color: graphics::Color::WHITE,
-                ..graphics::Text::default()
-            });
-        }
-
-        self.font.add(graphics::Text {
-            content: &(format!("{:.0}", progress.percentage()) + "%"),
-            position: graphics::Point::new(50.0, frame.height() / 2.0 + 50.0),
-            size: 30.0,
-            color: graphics::Color::WHITE,
-            ..graphics::Text::default()
-        });
-
-        self.font.draw(&mut frame.as_target());
     }
 }
