+++ src/state.rs
@@ -2,6 +2,11 @@ use crate::graphics::Window;
 use crate::load::Task;
 
 /// The state of your game.
+///
+/// Ideally, implementors of this trait should hold all the game data while
+/// having clear boundaries established by a powerful API.
+///
+/// [`State`]: trait.State.html
 pub trait State {
     /// Defines how many times the [`update`] function should be called per
     /// second.
@@ -11,16 +16,25 @@ pub trait State {
     /// [`update`]: #tymethod.update
     const TICKS_PER_SECOND: u16;
 
+    /// Loads the [`State`].
+    ///
+    /// Most of the time, you should simply use [`Task::new`] here.
+    ///
+    /// [`State`]: trait.State.html
+    /// [`Task::new`]: load/struct.Task.html#method.new
     fn load(window: &Window) -> Task<Self>
     where
         Self: Sized;
 
-    /// Update your game state here.
+    /// Updates the [`State`].
+    ///
+    /// All your game simulation logic should live here.
     ///
     /// The [`TICKS_PER_SECOND`] constant defines how many times this function
     /// will be called per second. This function may be called multiple times
     /// per frame if it is necessary.
     ///
+    /// [`State`]: trait.State.html
     /// [`TICKS_PER_SECOND`]: #associatedconstant.TICKS_PER_SECOND
     fn update(&mut self);
 }
